% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adr-package.R
\docType{package}
\name{adr-package}
\alias{adr}
\alias{adr-package}
\title{adr: Analiza Danych z R}
\description{
Dane do kursu "Analiza danych z R".
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mbojan/adr}
  \item Report bugs at \url{https://github.com/mbojan/adr/issues}
}

}
\author{
\strong{Maintainer}: Michal Bojanowski \email{michal2992@gmail.com} (\href{https://orcid.org/0000-0001-7503-852X}{ORCID}) (Kozminski University)

}
\keyword{internal}
