% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dekoduj.R
\encoding{UTF-8}
\name{dekoduj}
\alias{dekoduj}
\alias{dekoduj.data.frame}
\alias{dekoduj.default}
\alias{koduj}
\title{Koduj lub dekoduj napisy z UTF-8 do Windows lub vice versa}
\usage{
dekoduj(x, ...)

\method{dekoduj}{data.frame}(x, from_encoding = "UTF-8",
  to_encoding = "cp1250", ...)

\method{dekoduj}{default}(x, from_encoding = "UTF-8",
  to_encoding = "CP1250", ...)

koduj(x, from_encoding = "", to_encoding = "UTF-8", ...)
}
\arguments{
\item{x}{ramka, lub wektor napisów do (de)kodowania, patrz Details}

\item{...}{inne argumenty dla \code{\link{iconv}}

Funkcja \code{dekoduj} bierze ramkę danych z zwraca ją ze wszystkimi
kolumnami napisowymi (character) przekodowanymi z UTF-8 na kodowanie CP1250
(Windows po Polsku).

Funkcja \code{koduj} bierze wektor napisów w kodowaniu CP1250 i zwraca
przekształcony na UTF-8.}

\item{from_encoding, to_encoding}{strona kodowa źródłowa i docelowa}
}
\description{
Zmiana strony kodowej UTF-8 <-> CP1250. Przydatne w przypadku interakcji z
bazami danych takimi, jak SQLite, w których kwerendy muszą być wysyłane w
UTF-8. W tym samym kodowaniu zwracane są wyniki.
}
