% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kopiuj_dane.R
\encoding{UTF-8}
\name{kopiuj_dane}
\alias{kopiuj_dane}
\title{Kopiuj dane do lokalnego katalogu}
\usage{
kopiuj_dane(folder = "data", dryrun = FALSE)
}
\arguments{
\item{folder}{napis, katalog, do którego skopiować dane. Domyślnie 'data'.}

\item{dryrun}{czy wypisać "na sucho" operacje do wykonania}
}
\description{
W katalogu roboczym tworzy podkatalog 'data' i kopiuje tam wszystkie pliki
potrzebne podczas warsztatu.
}
