% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncn.R
\docType{data}
\encoding{UTF-8}
\name{ncn}
\alias{ncn}
\title{Statystyki składania wniosków do NCN}
\format{Ramka danych: 2979 obserwacje, 14 kolumn
\describe{
\item{GWO}{Grupa Wspólnej Oceny}
\item{jednostka_uczelniana}{Nazwa jednostki podstawowej (wydziały/instytuty)}
\item{jednostka_glowna}{Nazwa instytucji naukowej}
\item{zrodlo_danych}{Nazwa źródła danych o liczbie pracowników naukowych w jednostkach}
\item{grupa}{Grupa dziedzinowa, wtórna do GWO}
\item{rok}{}
\item{kategoria_naukowa_2015}{Kategoria naukowa jednostki za rok 2015}
\item{kwota}{Łączna kwota przyznanego dofinansowania z NCN}
\item{liczba_pracownikow}{Liczba pracowników naukowych}
\item{wnioski_finansowane}{Liczba złożonych wniosków, które otrzymały finansowanie}
\item{wnioski_zlozone}{Liczba złożonych wniosków}
\item{GWO_grupa,GWO_typ_jednostki,GWO_dziedzina}{Wydorębnione z \code{GWO} odpowiednio: grupa dziedzin nauki, typ jednostki głównej oraz dziedzina nauki}
}}
\source{
Narodowe Centrum Nauki http://www.ncn.gov.pl
}
\description{
Statystyki składania wniosków przez polskie instytucje
naukowe do Narodowego Centrum Nauki w latach 2013-15.
}
