% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pgss.R
\docType{data}
\encoding{UTF-8}
\name{pgss}
\alias{pgss}
\alias{pgss2}
\title{Polskie Generalne Sondaże Społeczne}
\format{
Ramka danych: 16234 obserwacje, 73 kolumny

Opisy kolumn oraz dostępność pytań w latach opisana jest w pliku
'pgss-info.xls' oraz w winiecie \href{../doc/pgss.html}{PGSS}
}
\source{
Archiwum Danych Społecznych http://www.ads.org.pl
}
\description{
Podzbiór Polskich Generalnych Sondaży Społecznych
}
\examples{
# Ścieżka do pliku XLS
system.file(file.path("exdata", "pgss-info.xls"), package="adr")
}
