% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uczniowie.R
\docType{data}
\encoding{UTF-8}
\name{uczniowie}
\alias{uczniowie}
\alias{uczniowie_relacje}
\alias{uczniowie_atrybuty}
\title{Dane o uczniach pewnej klasy podstawowej}
\format{
Dwie ramki danych z następującymi kolumnami:

\code{uczniowie_relacje} kto z kim
\describe{
\item{from}{ID nadawcy relacji}
\item{to}{ID odbiorcy relacji}
\item{type}{Rodzaj relacji: "play", "notplay", "secret", "work"}
}

\code{uczniowie_atrybuty} własności uczniów
\describe{
\item{name}{ID ucznia}
\item{isei08_m}{Status społeczno-ekonomiczny (ISEI) matki}
\item{isei08_f}{Status społeczno-ekonomiczny (ISEI) ojca}
\item{female}{Czy dziecko to dziewczynka}
}
}
\source{
Instyt Badań Edukacyjnych, badanie Szkolne Uwarunkowania Efektywności Kształcenia (SUEK)
}
\description{
Dane o 26-tu uczniach klasy podstawowej z pewnej szkoły w Polsce.
}
