% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Warszawa.R
\docType{data}
\encoding{UTF-8}
\name{Warszawa}
\alias{Warszawa}
\title{Dzielnice Warszawy}
\format{
Ramka danych z następującymi kolumnami:
\describe{
\item{Kod}{Kod TERYT}
\item{Dzielnica}{Nazwa dzielnicy}
\item{Strona_wisly}{Strona Wisły: lewa/prawa}
\item{ekologiczne}{Teren: użytki ekologiczne [ha]}
\item{inne}{Teren: inne użytki [ha]}
\item{las}{Teren: zalesiony lub zakrzewiony [ha]}
\item{nieużytki}{Teren: nieużytki [ha]}
\item{rolne}{Teren: rolniczy [ha]}
\item{woda}{Teren: woda (rzeki, stawy etc.) [ha]}
\item{zabudowane}{Teren: zabudowany [ha]}
\item{kobiety}{Ludność: liczba kobiet}
\item{mężczyźni}{LUdność: liczba mężczyzn}
\item{gestosc}{Gęstość zaludnienia [os./km^2]}
}
}
\source{
Bank Danych Lokalnych GUS http://stat.gov.pl/bdl/app/strona.html?p_name=indeks
}
\description{
Dane o 18-tu dzielnicach Warszawy z Banku Danych Lokalnych.
}
